Extracted Cookies from Storage
