+++ lib/api/src/cookies/events.ts
@@ -1,0 +1,19 @@
+import { Event } from '@exteranto/core'
+
+export class CookieChangedEvent extends Event {
+  /**
+   * @param {any} cookie
+   */
+  constructor (private cookie: any) {
+    super()
+  }
+
+  /**
+   * Cookie getter.
+   *
+   * @return {any}
+   */
+  public getCookie () : any {
+    return this.cookie
+  }
+}
