+++ lib/api/src/index.ts
@@ -1,41 +1,7 @@
-/**
- * Exporting Runttime APIs Driver.
- *
- */
-
-export * from './runtime'
-
-/**
- * Exporting Permissions APIs Driver.
- *
- */
-
-export * from './permissions'
-
-/**
- * Exporting BrowserAction APIs Driver.
- *
- */
-
 export * from './browserAction'
-
-/**
- * Exporting Messaging Driver.
- *
- */
-
+export * from './cookies'
 export * from './messaging'
-
-/**
- * Exporting Storage APIs Driver.
- *
- */
-
+export * from './permissions'
+export * from './runtime'
 export * from './storage'
-
-/**
- * Exporting Tabs APIs Driver.
- *
- */
-
 export * from './tabs'
