+++ lib/api/src/storage/events.ts
@@ -27,21 +27,3 @@ export class StorageChangedEvent extends Event {
     return this.storable
   }
 }
-
-export class CookieChangedEvent extends Event {
-  /**
-   * @param {any} cookie
-   */
-  constructor (private cookie: any) {
-    super()
-  }
-
-  /**
-   * Cookie getter.
-   *
-   * @return {any}
-   */
-  public getCookie () : any {
-    return this.cookie
-  }
-}
