+++ lib/api/src/storage/index.ts
@@ -1,7 +1,4 @@
 export { Storage } from './Storage'
 export { StorageProvider } from './StorageProvider'
 
-export { Cookies } from './Cookies'
-export { CookiesProvider } from './CookiesProvider'
-
 export * from './events'
