+++ lib/api/test/cookies/spec/ChromeCookies.spec.ts
@@ -1,10 +1,10 @@
 import { expect } from 'chai'
 import * as sinon from 'sinon'
-import * as browser from 'sinon-chrome/extensions'
+import * as chrome from 'sinon-chrome'
 import { Container } from '@exteranto/core'
 import { Browser } from '@exteranto/core'
 import { Dispatcher } from '@exteranto/core'
-import { CookieChangedEvent, Cookies } from '../../../../src'
+import { CookieChangedEvent, Cookies } from '../../../src'
 import {
   EmptyResponseException,
   InvalidCookieRequestException
@@ -13,79 +13,83 @@ import {
 declare var global: any
 
 export const tests = () => {
-  describe('Extensions', () => {
+  describe('Chrome', () => {
     let cookies
 
     before(() => {
-      Container.bindParam('browser', Browser.EXTENSIONS)
+      Container.bindParam('browser', Browser.CHROME)
 
       cookies = Container.resolve(Cookies)
     })
 
     afterEach(() => {
       Container.resolve(Dispatcher).events = {}
+      chrome.runtime.lastError = undefined
     })
 
     it('gets a cookie by name and url', async () => {
-      browser.cookies.get.resolves({ name: 'cookie' })
+      chrome.cookies.get.yields({ name: 'cookie' })
 
       expect(await cookies.get('url', 'cookie'))
         .to.have.property('name').which.is.equal('cookie')
 
-      sinon.assert.calledOnce(browser.cookies.get)
+      sinon.assert.calledOnce(chrome.cookies.get)
     })
 
     it('throws appropriate exception if cookie doesn\'t exist', async () => {
-      browser.cookies.get.resolves(null)
+      chrome.cookies.get.yields(null)
 
       await expect(cookies.get('url', 'cookie'))
         .to.eventually.be.rejectedWith(EmptyResponseException)
 
-      sinon.assert.calledOnce(browser.cookies.get)
+      sinon.assert.calledOnce(chrome.cookies.get)
     })
 
     it('throws appropriate exception if get method fails', async () => {
-      browser.cookies.get.rejects()
+      chrome.cookies.get.yields(null)
+      chrome.runtime.lastError = { message: 'failed' }
 
       await expect(cookies.get('url', 'cookie'))
         .to.eventually.be.rejectedWith(InvalidCookieRequestException)
 
-      sinon.assert.calledOnce(browser.cookies.get)
+      sinon.assert.calledOnce(chrome.cookies.get)
     })
 
     it('gets all cookies based on params', async () => {
-      browser.cookies.getAll.resolves([1, 2, 3])
+      chrome.cookies.getAll.yields([1, 2, 3])
 
       await expect(cookies.getAll({}))
         .to.eventually.be.an('array').which.has.lengthOf(3)
 
-      sinon.assert.calledOnce(browser.cookies.getAll)
+      sinon.assert.calledOnce(chrome.cookies.getAll)
     })
 
     it('throws appropriate exception if get all method fails', async () => {
-      browser.cookies.getAll.rejects()
+      chrome.cookies.getAll.yields(undefined)
+      chrome.runtime.lastError = { message: 'failed' }
 
       await expect(cookies.getAll({}))
         .to.eventually.be.rejectedWith(InvalidCookieRequestException)
 
-      sinon.assert.calledOnce(browser.cookies.getAll)
+      sinon.assert.calledOnce(chrome.cookies.getAll)
     })
 
     it('sets a cookie', async () => {
-      browser.cookies.set.resolves(undefined)
+      chrome.cookies.set.yields(undefined)
 
       await expect(cookies.set({})).not.to.be.eventually.rejected
 
-      sinon.assert.calledOnce(browser.cookies.set)
+      sinon.assert.calledOnce(chrome.cookies.set)
     })
 
     it('throws appropriate exception if set method fails', async () => {
-      browser.cookies.set.rejects()
+      chrome.cookies.set.yields(undefined)
+      chrome.runtime.lastError = { message: 'failed' }
 
       await expect(cookies.set({}))
         .to.eventually.be.rejectedWith(InvalidCookieRequestException)
 
-      sinon.assert.calledOnce(browser.cookies.set)
+      sinon.assert.calledOnce(chrome.cookies.set)
     })
 
     it('registers event listener', (done) => {
@@ -100,7 +104,7 @@ export const tests = () => {
           } catch (e) { done(e) }
         })
 
-      browser.cookies.onChanged.trigger('cookie')
+      chrome.cookies.onChanged.trigger('cookie')
     })
   })
 }
