+++ lib/api/test/cookies/spec/ExtensionsCookies.spec.ts
@@ -1,10 +1,10 @@
 import { expect } from 'chai'
 import * as sinon from 'sinon'
-import * as chrome from 'sinon-chrome'
+import * as browser from 'sinon-chrome/extensions'
 import { Container } from '@exteranto/core'
 import { Browser } from '@exteranto/core'
 import { Dispatcher } from '@exteranto/core'
-import { CookieChangedEvent, Cookies } from '../../../../src'
+import { CookieChangedEvent, Cookies } from '../../../src'
 import {
   EmptyResponseException,
   InvalidCookieRequestException
@@ -13,83 +13,79 @@ import {
 declare var global: any
 
 export const tests = () => {
-  describe('Chrome', () => {
+  describe('Extensions', () => {
     let cookies
 
     before(() => {
-      Container.bindParam('browser', Browser.CHROME)
+      Container.bindParam('browser', Browser.EXTENSIONS)
 
       cookies = Container.resolve(Cookies)
     })
 
     afterEach(() => {
       Container.resolve(Dispatcher).events = {}
-      chrome.runtime.lastError = undefined
     })
 
     it('gets a cookie by name and url', async () => {
-      chrome.cookies.get.yields({ name: 'cookie' })
+      browser.cookies.get.resolves({ name: 'cookie' })
 
       expect(await cookies.get('url', 'cookie'))
         .to.have.property('name').which.is.equal('cookie')
 
-      sinon.assert.calledOnce(chrome.cookies.get)
+      sinon.assert.calledOnce(browser.cookies.get)
     })
 
     it('throws appropriate exception if cookie doesn\'t exist', async () => {
-      chrome.cookies.get.yields(null)
+      browser.cookies.get.resolves(null)
 
       await expect(cookies.get('url', 'cookie'))
         .to.eventually.be.rejectedWith(EmptyResponseException)
 
-      sinon.assert.calledOnce(chrome.cookies.get)
+      sinon.assert.calledOnce(browser.cookies.get)
     })
 
     it('throws appropriate exception if get method fails', async () => {
-      chrome.cookies.get.yields(null)
-      chrome.runtime.lastError = { message: 'failed' }
+      browser.cookies.get.rejects()
 
       await expect(cookies.get('url', 'cookie'))
         .to.eventually.be.rejectedWith(InvalidCookieRequestException)
 
-      sinon.assert.calledOnce(chrome.cookies.get)
+      sinon.assert.calledOnce(browser.cookies.get)
     })
 
     it('gets all cookies based on params', async () => {
-      chrome.cookies.getAll.yields([1, 2, 3])
+      browser.cookies.getAll.resolves([1, 2, 3])
 
       await expect(cookies.getAll({}))
         .to.eventually.be.an('array').which.has.lengthOf(3)
 
-      sinon.assert.calledOnce(chrome.cookies.getAll)
+      sinon.assert.calledOnce(browser.cookies.getAll)
     })
 
     it('throws appropriate exception if get all method fails', async () => {
-      chrome.cookies.getAll.yields(undefined)
-      chrome.runtime.lastError = { message: 'failed' }
+      browser.cookies.getAll.rejects()
 
       await expect(cookies.getAll({}))
         .to.eventually.be.rejectedWith(InvalidCookieRequestException)
 
-      sinon.assert.calledOnce(chrome.cookies.getAll)
+      sinon.assert.calledOnce(browser.cookies.getAll)
     })
 
     it('sets a cookie', async () => {
-      chrome.cookies.set.yields(undefined)
+      browser.cookies.set.resolves(undefined)
 
       await expect(cookies.set({})).not.to.be.eventually.rejected
 
-      sinon.assert.calledOnce(chrome.cookies.set)
+      sinon.assert.calledOnce(browser.cookies.set)
     })
 
     it('throws appropriate exception if set method fails', async () => {
-      chrome.cookies.set.yields(undefined)
-      chrome.runtime.lastError = { message: 'failed' }
+      browser.cookies.set.rejects()
 
       await expect(cookies.set({}))
         .to.eventually.be.rejectedWith(InvalidCookieRequestException)
 
-      sinon.assert.calledOnce(chrome.cookies.set)
+      sinon.assert.calledOnce(browser.cookies.set)
     })
 
     it('registers event listener', (done) => {
@@ -104,7 +100,7 @@ export const tests = () => {
           } catch (e) { done(e) }
         })
 
-      chrome.cookies.onChanged.trigger('cookie')
+      browser.cookies.onChanged.trigger('cookie')
     })
   })
 }
