+++ lib/api/test/storage/spec/SafariStorage.spec.ts
@@ -2,7 +2,7 @@ import { assert, expect } from 'chai'
 import { Container } from '@exteranto/core'
 import { Browser } from '@exteranto/core'
 import { Dispatcher } from '@exteranto/core'
-import { StorageChangedEvent, Storage } from '../../../../src'
+import { StorageChangedEvent, Storage } from '../../../src'
 
 export const tests = () => {
   describe('Safari', async () => {
